+++ deltafi-ui/src/components/header/AppTopBar.vue
@@ -1,8 +1,8 @@
 <template>
-  <nav class="navbar navbar-dark bg-dark p-0 shadow">
+  <nav class="navbar navbar-dark bg-dark p-0 shadow" :style="{ backgroundColor: backgroundColor, color: textColor }">
     <div class="navbar-brand col mr-0 px-3">
       <div class="row">
-        <div class="col title">{{ uiConfig.title }}</div>
+        <div class="col title" :style="{ color: textColor }">{{ uiConfig.title }}</div>
         <div class="col text-right">
           <button v-tooltip.right="'Toggle sidebar menu'" class="navbar-toggler btn btn-sm" @click="toggleSidebar">
             <i :class="toggleSidebarIcon" />
@@ -24,6 +24,14 @@ import { computed, inject } from "vue";
 
 const uiConfig = inject('uiConfig');
 const sidebarHidden = inject('sidebarHidden');
+
+const textColor = computed(() => {
+  return (uiConfig.topBar.textColor) ? `${uiConfig.topBar.textColor} !important` : null;
+});
+
+const backgroundColor = computed(() => {
+  return (uiConfig.topBar.backgroundColor) ? `${uiConfig.topBar.backgroundColor} !important` : null;
+});
 
 const toggleSidebarIcon = computed(() => {
   return sidebarHidden.value ? "pi pi-angle-double-right" : "pi pi-angle-double-left";