+++ deltafi-ui/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Versions page now groups versions by Core, Plugins, and Other. (#147)
+- Made top bar color configurable. (#152)
 
 ### Fixed
 - Margin in content viewer. (#146)